BACKPORT: VARIANT 2: arm64: Allow reference symbols in alt-seqs to be in kernel-data.
When patching-in alternative sequences during bootup, we copy
the alt-seq instructions to one or more target locations.
This relocation requires a fixup of references to symbols,
if they are outside of the alt-seq.

The decision if a fixup is required is implemented in the function
branch_insn_requires_update(), which checks if the given address
is in kernel-text (e.g. functions).

This patch renames the function to address_needs_relocation_fixup()
and also supports symbols in kernel-data (e.g. global variables).
This allows adrp to address such symbols, which is required for
e.g. ARM64_MISMATCHED_CACHE_LINE_SIZE.

This patch is part of the eMAG Spectre variant 2 patch set and required for
non-eMAG system.
This patch is backported from:
https://git.theobroma-systems.com/ampere-computing/linux.git/commit/?h=linux-4.16.12-amp&id=cd70cc560c99d814e0d10e1198cac7a91c124821

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>